# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "金丝雀蜜罐 CanaryGuard AntiFraud",
    "description": "蜜罐反诈双形态主控：REST API 同一内核，WebUI/MCP 均为客户端（D1）。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/traps": {
      "get": {
        "tags": [
          "traps"
        ],
        "summary": "List Traps",
        "description": "蜜饵列表（可按状态过滤）。",
        "operationId": "list_traps_api_v1_traps_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Traps Api V1 Traps Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "traps"
        ],
        "summary": "Create Trap",
        "description": "创建蜜饵草稿（用户自拟文案，启发式伪装度）。写操作需 admin（P1-1）。",
        "operationId": "create_trap_api_v1_traps_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrapRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Trap Api V1 Traps Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traps/generate-draft": {
      "post": {
        "tags": [
          "traps"
        ],
        "summary": "Generate Draft",
        "description": "LLM/模板生成蜜饵草稿（HITL：只产文案+指纹，不发布）。写操作需 admin（P1-1）。\n\nR4-F6（队1 P2-10）：platform 限枚举（_ALLOWED_PLATFORMS），非法值 422 invalid_platform。",
        "operationId": "generate_draft_api_v1_traps_generate_draft_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Generate Draft Api V1 Traps Generate Draft Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traps/{trap_id}": {
      "get": {
        "tags": [
          "traps"
        ],
        "summary": "Get Trap",
        "operationId": "get_trap_api_v1_traps__trap_id__get",
        "parameters": [
          {
            "name": "trap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trap Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Trap Api V1 Traps  Trap Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traps/{trap_id}/deploy": {
      "post": {
        "tags": [
          "traps"
        ],
        "summary": "Deploy Trap",
        "description": "标记已部署（用户手动发布后显式调用，draft→active）。写操作需 admin（P1-1）。",
        "operationId": "deploy_trap_api_v1_traps__trap_id__deploy_post",
        "parameters": [
          {
            "name": "trap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trap Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DeployRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Deploy Trap Api V1 Traps  Trap Id  Deploy Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traps/{trap_id}/monitor": {
      "post": {
        "tags": [
          "traps"
        ],
        "summary": "Monitor Trap",
        "description": "进入监控（active→monitored）。写操作需 admin（P1-1）。",
        "operationId": "monitor_trap_api_v1_traps__trap_id__monitor_post",
        "parameters": [
          {
            "name": "trap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trap Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Monitor Trap Api V1 Traps  Trap Id  Monitor Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traps/{trap_id}/disable": {
      "post": {
        "tags": [
          "traps"
        ],
        "summary": "Disable Trap",
        "description": "停用（enabled=0，仅 active/monitored）。写操作需 admin（P1-1）。",
        "operationId": "disable_trap_api_v1_traps__trap_id__disable_post",
        "parameters": [
          {
            "name": "trap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trap Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Disable Trap Api V1 Traps  Trap Id  Disable Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traps/{trap_id}/retire": {
      "post": {
        "tags": [
          "traps"
        ],
        "summary": "Retire Trap",
        "description": "手动退役（draft/active/monitored/hit → retired）。\n\n敏感操作（§11 / P1-2）：confirm=true + reason≥20 双确认闸控，\n成功后写 gate_logs（action='trap.retire'，before/after/payload_hash）+ 事件。",
        "operationId": "retire_trap_api_v1_traps__trap_id__retire_post",
        "parameters": [
          {
            "name": "trap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trap Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RetireRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Retire Trap Api V1 Traps  Trap Id  Retire Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traps/{trap_id}/check-hit": {
      "post": {
        "tags": [
          "traps"
        ],
        "summary": "Check Hit",
        "description": "输入候选文本检测踩饵；命中即退役（hit→retired，实锤 D2）。\n\nP2-4：由 GET 改为 POST（带状态迁移副作用，不符合 GET 语义）；\n请求体 {text: ...}。原 GET 路径返回 405。\nP1-C（R2 严格验收）：命中后与 job_trap_recheck 同口径联动升级——按显式\ndetection_id 或 text 反查最近匹配检测，finalize(trap_hit_count=1) 升级 L5 +\n告警 + trap_hit_escalated 事件（source=manual_check_hit），消除\"手工实锤\n不进案例管道\"的残留面；未匹配到检测则返回 escalation.status=skipped，\n命中结果本身不受影响。",
        "operationId": "check_hit_api_v1_traps__trap_id__check_hit_post",
        "parameters": [
          {
            "name": "trap_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trap Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckHitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Check Hit Api V1 Traps  Trap Id  Check Hit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckHitRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Text",
            "description": "待检测的候选文本"
          },
          "detection_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Id",
            "description": "可选：指定联动升级的检测 id；缺省按 text 反查最近匹配检测（P1-C）"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CheckHitRequest"
      },
...
      "CreateTrapRequest": {
        "properties": {
          "bait_text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 4,
            "title": "Bait Text",
            "description": "蜜饵文案"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "bait_text"
        ],
        "title": "CreateTrapRequest"
      },
...
      "DeployRequest": {
        "properties": {
          "target_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Url",
            "description": "用户手动发布的目标 URL"
          }
        },
        "type": "object",
        "title": "DeployRequest"
      },
...
      "GenerateDraftRequest": {
        "properties": {
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id",
            "description": "离线模板 id（t_invest_flow 等）"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform",
            "description": "埋设位置提示（限枚举，见 _ALLOWED_PLATFORMS）",
            "default": "评论区"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "GenerateDraftRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RetireRequest": {
        "properties": {
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "操作理由（≥20 字）",
            "default": ""
          }
        },
        "type": "object",
        "title": "RetireRequest",
        "description": "退饵双确认（§11 敏感操作：confirm=true + reason≥20，GateLog 审计）。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}